Implement "agent" mode. (#129)
* Add fetchLocalCommands helper method.

Adds a helper method to the ComplianceFetcher class. It can be used to execute
commands locally and retrieve output. The output is formatted to look like it
was copied from a terminal so that it can be stored as evidence.

* Fix locker branch implementation.

This change allows the branch to be specified via `locker.branch` in the main
configuration file. A new branch is created if it doesn't exist on the remote.

* Make local evidence repository path configurable.

This change allows the local evidence repository path to be specified via
`locker.local_path` in the main configuration file.

* Implement agent mode.

This change implements agent mode. Any evidence created in agent mode will be
stored under the corresponding agent directory. Agents will cryptographically
sign any evidence they fetch. Signed evidence can be used in checks and is
automatically verified when loaded from the locker.

* Add documentation explaining how to verify encrypted evidence.

* Add file 'compliance/utils/fetch_local_commands' to manifest.

* Attempt to import any missing fetchers from include JSON file.

* Correctly set agent when loading evidence from the cache.

* Add 'locker.ignore_signatures' configuration.

Set 'locker.ignore_signatures' to 'true' to skip signature verification.

* Use correct evidence metadata when fetching previous revisions.

Ensure the 'evidence_dt' is specified when fetching evidence metadata.

* Fix agent class property return types.

Change following properties to correctly return a boolean type:
- compliance.agent.ComplianceAgent.signable
- compliance.agent.ComplianceAgent.verifiable

* Add 'locker.default_branch' configuration.

Defaults to 'master'.

* Add 'locker.force_push' configuration.

Defaults to False.

* Add 'locker.agent_public_key' configuration.

Allow an Agent public key to be specified in the configuration.

* Update CHANGES.md for 1.22.0 release.