Remove the need of a --creds-path
If --creds-path is not provided it will not default to `~/.credentials` any more.
Instead, it will assume that env vars want to be used and it will try so.

Also, the documentation has been updated to reflect this and provide a way to use 1Password CLI as a credentials loader.